requirements.txt
alembic==1.12.1
asttokens==2.4.1
backcall==0.2.0
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
decorator==5.1.1
executing==2.0.0
flash==1.0.3
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
gunicorn==21.2.0
idna==3.4
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jwt==1.3.1
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.39
pure-eval==0.2.2
pycparser==2.21
pydotenv==0.0.7
Pygments==2.16.1
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.22
stack-data==0.6.3
traitlets==5.12.0
typing_extensions==4.8.0
urllib3==2.0.6
wcwidth==0.2.8
Werkzeug==3.0.0
WTForms==3.1.0
